2021/requirements.txt
pandas>=2.0
numpy>=1.26
PyYAML>=6.0
